#153 - Move conversion function.
