+++ src/nimble_model_data.h
@@ -158,22 +158,5 @@ protected:
 
 } // namespace nimble
 
-///
-/// Temporary Solution while refactoring
-///
-namespace details {
-
-nimble::ModelData& to_ModelData(const std::shared_ptr<nimble::ModelDataBase>& mptr)
-{
-  auto *model_data_ptr = dynamic_cast<nimble::ModelData*>(mptr.get());
-  if (model_data_ptr == nullptr) {
-    throw std::runtime_error(" Incompatible Model Data \n");
-  }
-  return *model_data_ptr;
-}
-/////////////////
-
-}
-///////////
 
 #endif
