+++ src/nimble_model_data_utils.cc
@@ -1,0 +1,77 @@
+/*
+//@HEADER
+// ************************************************************************
+//
+//                                NimbleSM
+//                             Copyright 2018
+//   National Technology & Engineering Solutions of Sandia, LLC (NTESS)
+//
+// Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government
+// retains certain rights in this software.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are
+// met:
+//
+// 1. Redistributions of source code must retain the above copyright
+// notice, this list of conditions and the following disclaimer.
+//
+// 2. Redistributions in binary form must reproduce the above copyright
+// notice, this list of conditions and the following disclaimer in the
+// documentation and/or other materials provided with the distribution.
+//
+// 3. Neither the name of the Corporation nor the names of the
+// contributors may be used to endorse or promote products derived from
+// this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY NTESS "AS IS" AND ANY EXPRESS OR IMPLIED
+// WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+// MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN
+// NO EVENT SHALL NTESS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+// TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+// Questions?  Contact David Littlewood (djlittl@sandia.gov)
+//
+// ************************************************************************
+//@HEADER
+*/
+
+#include "nimble_model_data_utils.h"
+#include "nimble_model_data.h"
+
+namespace nimble { namespace details {
+
+nimble::ModelData& to_ModelData(const std::shared_ptr<nimble::ModelDataBase>& mptr)
+{
+  auto *model_data_ptr = dynamic_cast<nimble::ModelData*>(mptr.get());
+  if (model_data_ptr == nullptr) {
+    throw std::runtime_error(" Incompatible Model Data \n");
+  }
+  return *model_data_ptr;
+}
+
+}}
+
+#ifdef NIMBLE_HAVE_KOKKOS
+
+#include "nimble_kokkos_model_data.h"
+
+namespace nimble_kokkos { namespace details_kokkos {
+
+nimble_kokkos::ModelData& to_ModelData(const std::shared_ptr<nimble::ModelDataBase>& mptr)
+{
+  auto *model_data_ptr = dynamic_cast<nimble_kokkos::ModelData*>(mptr.get());
+  if (model_data_ptr == nullptr) {
+    throw std::runtime_error(" Incompatible Model Data \n");
+  }
+  return *model_data_ptr;
+}
+
+}}
+
+#endif
