+++ src/nimble_mpi.cc
@@ -56,6 +56,7 @@
 #include "nimble_block_material_interface_factory_base.h"
 #include "nimble_material_factory.h"
 #include "nimble_model_data.h"
+#include "nimble_model_data_utils.h"
 
 #include "nimble_mpi.h"
 #include "nimble_vector_communicator.h"
